+++ src/components/MultiSeriesBarChart/types.ts
@@ -34,7 +34,15 @@ export interface AccessibilitySeries {
   }[];
 }
 
+export enum BarMargin {
+  Small = 0.1,
+  Medium = 0.3,
+  Large = 0.5,
+  None = 0,
+}
+
 export interface BarOptions {
+  margin: keyof typeof BarMargin;
   hasRoundedCorners: boolean;
   isStacked: boolean;
 }